Changes how company-backends are set per mode...
Each major mode should define a variable <mode-name>-company-backends
which should be initialized to whatever value is appropriate for that
mode. Each major mode should also define a company-setup-function
which is added to the mode-hook. This setup function is responsible
for setting the buffer-local value of company-backends to
<mode-name>-company-backends. In this way, new modules which wish to
register themselves as a valid backend for a given mode, can do so
simply by adding themselves to the list of
<mode-name>-company-backends.