Change project-treemacs-try logic
Whereas before we would only match a treemacs project if the current buffer
matched within a treemacs project/workspace, now we additionally fall-back to
matching the first treemacs project, regardless of where we make the call.

In practice this means that anytime the treemacs sidebar is visible, project.el
will prefer the project-treemacs backend.